requirements.txt
streamlit
pytesseract
Pillow
numpy
scikit-learn
PyPDF2
python-docx
